Add implementation of Disclosure - Notifications on bottom of screen
This diff adds a bottom notification.
It checks if the view controller responsible for call/present the notification is a tab bar controller so it can be added extra space at the bottom to avoid the notification displays over the tab bar.
If the user wants that behaviour, it is just necessary to use another view controller like a UINavigationController.

The 'guideline' used for this diff was the code on ShoutFactory.swift.