Update Demo project
This diff updates the demo project for bottom notification.
In the main view (ViewController) it updates with a new button for displaying the disclosure notification.
Also adds a new modal view which has a tab bar so it can be seen in action the disclosure appearing above the tab bar.